requirements.txt
gitdb==4.0.11
GitPython==3.1.43
lizard==1.17.10
PyDriller==2.6
python-markdown-generator==0.1.0
pytz==2024.1
smmap==5.0.1
types-pytz==2024.1.0.20240417
